requirements/base.txt
pytz
python-slugify
Pillow
argon2-cffi
whitenoise
redis
hiredis
celery
django-celery-beat
flower

# Images management
# ------------------------------------------------------------------------------
imagekitio
cloudinary

# Parsing
# ------------------------------------------------------------------------------
beautifulsoup4

# Socialmedias
# ------------------------------------------------------------------------------
tweepy

# Finance data
# ------------------------------------------------------------------------------
yahooquery
yfinance

# Django
# ------------------------------------------------------------------------------
django
django-cors-headers
django-environ
gunicorn
django-allauth
django-crispy-forms
crispy-bootstrap5
django-redis
# Django REST Framework
djangorestframework

# Django extras
# ------------------------------------------------------------------------------
django-widget-tweaks
django-ckeditor
django-cleanup
django-countries
django-admin-honeypot
django-import-export
geoip2
psycopg2-binary # https://github.com/psycopg/psycopg2
pandas
html5lib

# Payment
# ------------------------------------------------------------------------------
stripe